requirements.txt
# Core Data Science & ML
jupyter==1.0.0
notebook==7.0.7
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1

# AI/LLM APIs
openai>=1.0.0
anthropic==0.8.1
langchain==0.1.0
transformers==4.36.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
pytest==7.4.4

# Vector Databases (for storing embeddings)
chromadb==0.4.22

# Experiment Tracking
wandb==0.16.2

# Together AI
together 

# requests
requests==2.31.0

# PyPDF2
PyPDF2==3.0.2

